Multiple PR fixes - reorg/rename files, casing, use undefined,
