+++ packages/iobio-react-components/src/utils/iobioTypes.ts
@@ -25,6 +25,7 @@ export type DataUpdate = StatisticsData & PercentData;
 export type HistogramData = { [K in BamHistogramKey]: { [numKey: string]: number } };
 export type StatisticsData = { [K in StatisticKey]: number };
 export type PercentData = { [K in BamPercentKey]: number };
+export type IobioMetaData = { [k in BamKey]: number };
 export type StatsOutput = {
-	iobio_metadata: { [k in BamKey]: number };
+	iobio_metadata: IobioMetaData;
 };
